Revert "Override `c_compiler` on Windows"
This reverts commit 50f506d7a9eda11f56c4ff8083b0bfcf82a874ca.

Since this workaround was added, `conda-build`'s behavior has been fixed
to handle selectors on `compiler`s. Given this, the workaround made in
commit is no longer needed and can be reverted. It's likely this
workaround was not totally sufficient given we were still having issues
with Python versions being dropped in the re-render.