MNT: Re-rendered with conda-smithy 3.1.11 and pinning 2018.08.05
Now that the selectors have been changed to Python 2.7, re-render to
update compiler variants. This still doesn't look completely right
unfortunately as `vs` and `m2w64` compilers are both being included in
both versions of the build. Perhaps the build will be correct though.

Note: Interestingly this fails to change anything during the re-render
some of the time.