Skip on Python 3.3 onwards (where part of Python)
This is a backport of Python 3.3's 'lzma' module for
XZ/LZMA compressed files.

I can only think of two reasons to install the back port
on Python 3.3 or later:

First, if your Python was compiled without lzma, which
should not happen with conda.

Second, for testing for a difference in behaviour between
the backport and the upstream library. The later is a
corner case, and to me not worth the pain in getting this
recipe to work on recent compilers.